# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Forum App",
        "contact": {
            "email": "rizkimahjati845@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/forum/api",
            "description": "API Forum App"
        }
    ],
    "paths": {
        "/auth/checkLogin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authentication - Check Login",
                "description": "Authentication User",
                "operationId": "checkLogin",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "username": "admin",
                                "password": "admin"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil Login",
                                    "data": [
                                        {
                                            "username": "admin",
                                            "nama": "Admin",
                                            "role": "Admin"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authentication - Register",
                "description": "Post data Register",
                "operationId": "register",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "username": "admin",
                                "password": "admin",
                                "nama": "Admin",
                                "alamat": "Pemalang",
                                "no_telp": "087723977966",
                                "jenis_kelamin": "Perempuan",
                                "role": "Admin"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil disimpan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    }
}